requirements.txt

# ===== Core Framework =====
fastapi==0.104.0
uvicorn[standard]==0.24.0

# ===== Data Validation & Settings =====
pydantic==2.4.0
pydantic-settings==2.0.0
email-validator==2.1.0

# ===== Database & ORM =====
sqlalchemy==2.0.0
alembic==1.12.0
asyncpg==0.28.0

# ===== Authentication & Security =====
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# ===== AI Integration =====
google-generativeai==0.3.0

# ===== File Storage =====
boto3==1.28.0
botocore==1.31.0

# ===== Cache & Background Tasks =====
redis==5.0.0
celery==5.3.0

# ===== HTTP Client =====
httpx==0.25.0
aiofiles==23.2.0

# ===== Utilities =====
python-dotenv==1.0.0
python-slugify==8.0.1

# ===== Email (Optional) =====
emails==0.6

# ===== Monitoring & Logging =====
sentry-sdk[fastapi]==1.32.0

# ===== Development Dependencies =====
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
pytest-xdist==3.3.0
factory-boy==3.3.0
faker==19.6.0

# ===== Code Quality =====
black==23.9.0
isort==5.12.0
flake8==6.0.0
mypy==1.5.0
pre-commit==3.4.0

# ===== Documentation =====
mkdocs==1.5.3
mkdocs-material==9.2.0

# ===== Load Testing =====
locust==2.17.0